backend: import emitter trait in tauri modules
