+++ Backend/src/tauri_commands/general.rs
@@ -1,0 +1,27 @@
+use tauri::{Runtime, Window};
+
+use crate::utilities::utilities;
+
+#[tauri::command]
+pub fn about_info() -> utilities::AboutInfo {
+    utilities::AboutInfo::gather()
+}
+
+#[tauri::command]
+pub fn show_licenses() -> Result<(), String> {
+    // open a bundled licenses file, or a window, or external URL
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn browse_directory<R: Runtime>(
+    window: Window<R>,
+    purpose: Option<String>,
+) -> Option<String> {
+    let title = match purpose.as_deref() {
+        Some("clone_dest") => "Choose destination folder",
+        Some("add_repo") => "Select an existing Git repository folder",
+        _ => "Select a folder",
+    };
+    utilities::browse_directory_async(window.app_handle().clone(), title).await
+}
